requirements.txt
tk
python-can